Update Analyzer_Keras
This is comparing two genres (Classic Doom and Cavern Death). I'll be editing this to test it on each song and see how it does overall. From what I've tested so far, this works decently and seems to meet my goal baseline of 60% accuracy. Seems to get it completely wrong sometimes.